intelligence/requirements.txt
fastapi
uvicorn
pydantic
kafka-python
numpy
pandas
scikit-learn
joblib